% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/atlasmeta.R
\docType{data}
\name{atlasmeta}
\alias{atlasmeta}
\title{DGSA meta data}
\format{a large data.frame:
\describe{
 \item{id} {id field}
 \item{element} {name of the element, this is the column name in the atlas data.frame}
 \item{unit} {unit of measurement}
 \item{label} {a more descriptive label}
}}
\usage{
atlasmeta
}
\description{
Meta data of the Dutch Geochemical Soil Atlas.
}
\details{
This is the meta data of the Dutch Geochemical Soil Atlas. It
contains the field names of the elements in the \code{atlas}
data.frame, the units of measurement and a more descriptive label
}
\keyword{datasets}
