% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{dgsa}
\alias{dgsa}
\title{DGSA main function}
\usage{
dgsa(.clusters, .X, .normalize = TRUE, .nBoot = 100,
  .interactions = FALSE, .nBins = 3, .alpha = 0.95, .parallel = FALSE,
  .progress = TRUE)
}
\arguments{
\item{.normalize}{- Whether to normalize with .alpha boostrapped or not.}

\item{.nBoot}{- The number of bootstrapped samples for normalization computation.}

\item{.interactions}{- boolean, specifying whether to compute interactions (default = FALSE)}

\item{clusters}{- from some method like kmedoids}

\item{parameters}{- a data frame with parameters stacked in columns and designs in rows.}
}
\description{
DGSA main function
}
\details{
This code implements the original DGSA code by Celine Scheidt.
}
\author{
Ogy Grujic (ogyg@stanford.edu)
}

