% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distribution.R
\name{distribution}
\alias{distribution}
\title{Distribution Plot.}
\usage{
distribution(x, xbin = 10, xadj = 1)
}
\arguments{
\item{x}{name of the element}

\item{xbin}{bin size, see ?geom_histogram}
}
\description{
Create a distribution plot (histogram with density)
}
\details{
A distribution plot with density gives a  visual overview of the
statistical distribution of the data.
}
\examples{
p <- distribution("xrfal2o3")
print(p)

}
