% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/label2elem.R
\name{label2elem}
\alias{label2elem}
\title{translate label to element}
\usage{
label2elem(label)
}
\arguments{
\item{label}{Label to translate, case sensitive}
}
\value{
field name in database
}
\description{
Translate a full name of a component into the element name (field
name) of the dgsa database
}
\details{
The DGSA database is a table with many elements. Each field has
a name consiting of analytical technique and name of the element.
This functions translates the full name of an element (e.g. Al
Total) into the name of the field (e.g. xrfal2o3).
}
\examples{
label2elem("Al total")
# returns "xrfal2o3"

}
