% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map.R
\name{map}
\alias{map}
\title{Create a bubble map}
\usage{
map(x)
}
\arguments{
\item{x}{name of the element}
}
\description{
Create a spatial bubble map
}
\details{
A bubble map is a simple way to both show spatial distribution and
magnitude of measurements. It shows the data without an underlying
interpolation model like inverse distance and kriging.

In ths implementation ggplot2 is used to draw the map and the
bubbles (circles). Usually, a specific function is chosen to relate
the magnitude of the measurement to the size of the bubble, using
radius or area. The \code{map} function uses the \code{size}
argument in the \code{aes} function.
}
\examples{
p <- map("xrfal2o3")
print(p)

}
