% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{plotCDFS}
\alias{plotCDFS}
\title{plots CDF's}
\usage{
plotCDFS(.clustering, .X, .code = NULL, .nBins = 3, .ggReturn = "plot",
  lwd = 1)
}
\arguments{
\item{.X}{- A matrix of input parameters}

\item{.ggReturn}{- whether or not to return a ggplot structure (for adjustments)}

\item{.clusters}{- A vector of cluster tags.}

\item{.interaction}{- specify which interaction to plot i.e. "a|b", both "a" and "b"
have to be parameters names already present in the dataset.}
}
\details{
This code uses ggplot for plotting it will throw out a very basic plot
}
\author{
Ogy Grujic (\email{ogyg@stanford.edu})
}

