% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{plotDGSAscatter}
\alias{plotDGSAscatter}
\title{plots scaled scatter plots of inputs on provided low dimmensional representation}
\usage{
plotDGSAscatter(.inputs, .coords, .clusters = NULL, .ggReturn = FALSE,
  .cb = FALSE)
}
\arguments{
\item{.inputs}{- A data frame with input parameters stacked in columns}

\item{.coords}{- Coordinates matrix (only the first two columns matter), from cmdscale or pca analysis}

\item{.clusters}{- Optional clustering vector that puts different shape on different cluster}

\item{.ggReturn}{- Whether to return ggplot structure or not.}

\item{.cb}{- Whether to plot with colour blind friendly palette (default = FALSE)}
}
\details{
It plots a ggplot with facet wrap of all input parameters.
}
\author{
Ogy Grujic (\email{ogyg@stanford.edu})
}

