% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{plotMatrixDGSA}
\alias{plotMatrixDGSA}
\title{DGSA visualization routine}
\usage{
plotMatrixDGSA(.dgsa, .hypothesis = TRUE, .method = "circle", ...)
}
\arguments{
\item{.dgsa}{- a "DGSAstructure" class object as produced by \code{dgsa} function}

\item{.hypothesis}{- whether to mark sensitive or not sensitive (star = sensitive)}

\item{.method}{- method for corrPlot (circle is default);}

\item{...}{- whatever else setting you want to pass to corrPlot (see help(corrPlot))}
}
\details{
It simply computes the mean of the sensitivity matrix and passess it to
          corrPlot to visualize both interactions and main effects.
}
\author{
Ogy Grujic (\email{ogyg@stanford.edu})
}

