% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{plotParetoDGSA}
\alias{plotParetoDGSA}
\title{creates ggplot based pareto plots for dgsa}
\usage{
plotParetoDGSA(.dgsa, .clusters = FALSE, .interaction = NULL,
  .hypothesis = TRUE, .ggReturn = FALSE)
}
\arguments{
\item{.dgsa}{-  a "DGSAstructure" class object as produced by \code{dgsa} function}

\item{.clusters}{- whether to plot cluster specific L1 norms or not}

\item{.interaction}{- specify which interaction to plot i.e. "a|b", both "a" and "b"
have to be parameters names already present in the dataset.}

\item{.ggReturn}{- whether or not to return a ggplot structure (for adjustments)}
}
\details{
This code uses ggplot for plotting it will throw out a very basic plot
}
\author{
Ogy Grujic (\email{ogyg@stanford.edu})
}

