% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{scaleInp}
\alias{scaleInp}
\title{Scales input parameters to 0,1 scale}
\usage{
scaleInp(.Input, .lB = NULL, .uB = NULL)
}
\arguments{
\item{.Input}{- Input vector to scale}

\item{.lB}{- Lower bound if other than minimum}

\item{.uB}{- Upper bound if other than maximum}
}
\author{
Ogy Grujic (\email{ogyg@stanford.edu})
}

